Integrated Products and added new fields.

+++ order-service/src/main/java/com/order/service/mapper/OrderRowMapper.java
@@ -3,6 +3,7 @@ package com.order.service.mapper;
 import com.order.service.entities.Order;
 import io.r2dbc.spi.Row;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.function.BiFunction;
 
@@ -16,7 +17,7 @@ public class OrderRowMapper implements BiFunction<Row, Object, Order> {
         order.setId(row.get("id", String.class));
         order.setUserId(row.get("user_id", String.class));
         order.setStatus(row.get("status", String.class));
-        order.setAmount(row.get("amount", Double.class));
+        order.setAmount(row.get("amount", BigDecimal.class));
         order.setCreatedAt(row.get("created_at", LocalDateTime.class));
 
         return order;
